pyproject.toml
[tool.poetry]
name = "drf_api"
version = "0.1.0"
description = ""
authors = ["mohammedghafri <eng.m.ghafri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-admin = "^2.0.1"
django = "^3.1.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
